+++ app/src/main/java/com/realityexpander/tasky/presentation/login_screen/LoginEvent.kt
@@ -1,7 +1,7 @@
 package com.realityexpander.tasky.presentation.login_screen
 
 import com.realityexpander.tasky.domain.AuthInfo
-import com.realityexpander.tasky.ui.util.UiText
+import com.realityexpander.tasky.presentation.util.UiText
 
 sealed class LoginEvent {
     data class UpdateEmail(val email: String) : LoginEvent()
