+++ app/src/main/java/com/realityexpander/tasky/presentation/register_screen/RegisterScreen.kt
@@ -39,12 +39,12 @@ import com.realityexpander.tasky.domain.validation.ValidateUsername
 import com.realityexpander.tasky.domain.validation.validateEmail.ValidateEmailRegexImpl
 import com.realityexpander.tasky.presentation.destinations.LoginScreenDestination
 import com.realityexpander.tasky.presentation.login_screen.LoginScreenContent
-import com.realityexpander.tasky.ui.components.EmailField
-import com.realityexpander.tasky.ui.components.NameField
-import com.realityexpander.tasky.ui.components.PasswordField
+import com.realityexpander.tasky.presentation.components.EmailField
+import com.realityexpander.tasky.presentation.components.NameField
+import com.realityexpander.tasky.presentation.components.PasswordField
 import com.realityexpander.tasky.ui.theme.TaskyTheme
 import com.realityexpander.tasky.ui.theme.modifiers.*
-import com.realityexpander.tasky.ui.util.UiText
+import com.realityexpander.tasky.presentation.util.UiText
 
 @Composable
 @Destination
