+++ app/src/main/java/com/realityexpander/tasky/presentation/util/UiText.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.ui.util
+package com.realityexpander.tasky.presentation.util
 
 import android.content.Context
 import android.os.Parcelable
